requirements.txt
streamlit==1.37.0
pandas==2.1.3
numpy==1.26.2
pillow==10.3.0
pydub==0.25.1
soundfile==0.12.1
SpeechRecognition==3.10.0
python-dotenv==1.0.0
torch==2.7.0
torchaudio==2.2.0
transformers==4.50.0
ffmpeg-python==0.2.0
librosa==0.10.1
audio-recorder-streamlit==0.0.8
accelerate==1.7.0
# bitsandbytes is optional and platform-specific
# On CUDA systems, install with: pip install bitsandbytes